pyproject.toml
[tool.poetry]
name = "mini-framework"
version = "0.0.1"
description = "This is a mini framework. THIS ONLY FOR EDUCATIONAL PURPOSES."
authors = ["Vladyslav Timofeev <Timofeevvladyslav49@gmail.com>"]
readme = "README.md"
packages = [{include = "mini_framework", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
multidict = "^6.0.5"
pydantic = "^2.6.2"
jinja2 = { version = "^3.1.3", optional = true }
python-multipart = { version = "^0.0.9", optional = true }

[tool.poetry.extras]
templates = ["jinja2"]
multipart = ["python-multipart"]

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-random-order = "^1.1.1"
pytest-repeat = "^0.9.3"
httpx = "^0.26.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
pre-commit = "^3.6.0"

[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pyright]
reportMissingImports = false

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"